Add HuggingFaceData superclass
